pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "darkfield"
version = "0.1.0"
dependencies = [
  "numpy",
  "scipy",
  "pandas",
  "matplotlib",
  "opencv-python",
  "tqdm",
  "pyyaml",
  "astropy",
  "LightPipes",
]
requires-python = ">=3.12"
description = "Calculate darkfield."
readme = "README.md"



